backend/pyproject.toml
[project]
name = "archaeo-pro-backend"
version = "0.1.0"
description = "Backend for archaeo-pro — Italian archaeological surveillance reports."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "geoalchemy2>=0.15",
    "psycopg[binary]>=3.2",
    "alembic>=1.13",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-multipart>=0.0.12",
    "python-docx>=1.1",
    "docxtpl>=0.19",
    "httpx>=0.27",
    "pillow>=11.0",
    "shapely>=2.0",
    "geojson>=3.1",
    "exif>=1.6",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
